Replace Almost All Textures
This version replaces almost all textures with ones cloned directly from the Toontown Online branch of Project: Bikehorn instead of from the compressed download file which is where I foolishly got them from which that and my assumption pressing the download button on a different branch would download things from that branch at the time is what lead to TTR's textures being in here (Sorry about that! We're gonna work on making sure our assets and code are our own) as well as some of the original Open-Toontown resource textures.

This isn't being pushed to the main branch because there are some textures that are missing from the files that the game needs to function properly.