More Toon News for the Amused!
This update adds more images to the "Toon News for the Amused!" section which brings the remaining number of missing items down to 94 of the 420 items that I know about. This update also fixes the broken logo link, updates the old Arbitrium Studios Discord invite link, and moves the credits stuff to the Arbitrium Studios & Toontown Fantasy websites.